bug fixes and testing with postman

+++ src/utils/cloudinary.js
@@ -1,26 +1,35 @@
-import {v2 as cloudinary} from "cloudinary";
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from 'dotenv';
 
-cloudinary.config({ 
+// Config
+dotenv.config();
+
+cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY, 
-    api_secret: process.env.CLOUDINARY_API_SECRET 
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath) return null;
+        if (!localFilePath) return null;
 
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         })
 
         console.log("File is uploaded on Cloudinary", response.url);
+        fs.unlinkSync(localFilePath);
         return response;
-    } catch(error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved file from server is operation fails
+    } catch (error) {
+        console.error("Cloudinary Upload Error:", error);   // log the real error
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
         return null;
     }
 }
 
-export {uploadOnCloudinary};
+export { uploadOnCloudinary };
